Only_Images/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pillow==10.1.0
easyocr==1.7.1
transformers==4.36.2
torch==2.1.2
opencv-python==4.8.1.78
numpy==1.24.4
python-multipart==0.0.6
python-magic==0.4.27